+++ src/index.js
@@ -1,59 +1,7 @@
 import _ from 'lodash';
 import './style.css';
-
-class scores{
-  constructor(Name,score) {
-      this.Name=Name
-      this.score=score
-  }
-}
-const createNewgame=async() => {
-  const New_game = {
-      name: 'Luzinda games',
-    };
-    const key=localStorage.getItem("key");
-    if(key){
-        console.log(key)
-      return key
-    }
- const res = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(New_game),
-    });
-  console.log(res)
-  const data = await res.json();
-  console.log()
-  localStorage.setItem("key",data.result.substr(14, 20))
-  return result.substr(14, 20)
-}
-
-
-const superkey=createNewgame()
-
-const postNew=async(name,score)=>{
-  const data={
-  "user": name,
-  "score": score
-  }
-  console.log(data)
-  const keys=localStorage.getItem("key");
-  const res = await fetch(`https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/${keys}/scores/`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data),
-    });
-  const response=await res.json();
-  const li=document.createElement("li")
-  li.innerHTML=`${name} : ${score}`
-  const scores=document.querySelector(".scores").append(li)
-  
-}
-
+import {createNewgame,postNew} from "./modules/utils"
+createNewgame()
 const addItems=()=>{
 const submit = document.querySelector(".form-submit")
 submit.addEventListener("submit",(e)=>{
